+++ src/decode.h
@@ -12,7 +12,7 @@ typedef struct {
 
 float decode_transducer(const_scrappie_matrix logpost, float stay_pen, float skip_pen,
                         float local_pen, int *seq, bool allow_slip);
-char *overlapper(const int *seq, int n, int nkmer, int *pos);
+char *overlapper(const int *seq, size_t n, int nkmer, int *pos);
 char *homopolymer_dwell_correction(const event_table et, const int *seq,
                                    size_t nstate, size_t basecall_len);
 char *dwell_corrected_overlapper(const int *seq, const int *dwell, int n,
@@ -22,7 +22,7 @@ float sloika_viterbi(const_scrappie_matrix logpost, float stay_pen, float skip_p
                      float local_pen, int *seq);
 
 float argmax_decoder(const_scrappie_matrix logpost, int *seq);
-char *ctc_remove_stays_and_repeats(const int *seq, int n, int *pos);
+char *ctc_remove_stays_and_repeats(const int *seq, size_t n, int *pos);
 
 float decode_crf(const_scrappie_matrix trans, int * path);
 scrappie_matrix posterior_crf(const_scrappie_matrix trans);
@@ -34,17 +34,17 @@ float squiggle_match_viterbi(const raw_table signal, const_scrappie_matrix param
 
 
 
-bool are_bounds_sane(int const * low, int const * high, size_t nblock, size_t seqlen);
+bool are_bounds_sane(size_t const * low, size_t const * high, size_t nblock, size_t seqlen);
 float map_to_sequence_viterbi(const_scrappie_matrix logpost, float stay_pen, float skip_pen,
                               float local_pen, int const *seq, size_t seqlen, int * path);
 float map_to_sequence_forward(const_scrappie_matrix logpost, float stay_pen, float skip_pen,
                               float local_pen, int const *seq, size_t seqlen);
 float map_to_sequence_viterbi_banded(const_scrappie_matrix logpost, float stay_pen, float skip_pen,
                                      float local_pen, int const *seq, size_t seqlen,
-                                     int const * poslow, int const * poshigh);
+                                     size_t const * poslow, size_t const * poshigh);
 float map_to_sequence_forward_banded(const_scrappie_matrix logpost, float stay_pen, float skip_pen,
                                      float local_pen, int const *seq, size_t seqlen,
-                                     int const * poslow, int const * poshigh);
+                                     size_t const * poslow, size_t const * poshigh);
 
 
 #endif                          /* DECODE_H */
